Updating the Score Template files
Updating the score template files to provide the same type of nested
supported as with judges so that an event coordinator can create a score
template from within an event. Updating the view and support for a score
template as well. More to come for user story numbers 6, 8 & 9 as well
as adding questions to these score templates from within the same page
(user story 8 and 9)