Functionality implemented within competitors files
Increased functionality and support for the competitors model so that an
event coordinator can now create competitors from within a specific
event and allow them to be created with the hidden event_id field so
that they are associated with a specific event. This commit corresponds
to user stories 10 and 11 (competitor support).