Event Coordinator functionality
This commit updates the event_coordinator view for the "form" and "show"
pages as well as the coordinator_sessions helper. There is reformatting
of these views so that they match the theme/view supported within the
other models. This corresponds to an event_coordinator wanting to view
their profile and navigate through the pages (user stories 1, 6, 7, 8).