Event views and functionality
Increased functionality for the events pages so that the event
coordinator can now have the options to edit the event in such a way
that allows them to click the link to create judges, score templates,
and competitors all from within one page. This is the commit that wraps
together all of the previous commits from today and allows the
functionality to work together from the event coordinators perspective.
User stories 6, 7, 8, 9, 10 and 11 were all thought of in terms of tying
these commits and functionality together. I also added the "capybara"
gem to the Gemfile for testing purposes for the RSpec tests as that is
part of the problem I was having with the 'visit' method previously not
working. Still having a problem with testing the pages, however.