Updating competitor and template models
Configuring the model with a scoped "for_event" lambda function so that
the comptetiors and score template will be viewable within the event
show page. This functionality allows for a particular event page to show
the competitors and score template for just that event. This commit
corresponds to user stories 8 and 10.