Refactoring views to showcase new functionality
Updating the view pages according to the previous commits that allowed
for competitors, templates, and judges to all be shown for a specific
event. This commit allows for the events "show" page to render specific
partials that show all of that events competitors, judges, and score
templates. More functionality to come within each of these collapsible
objects. This commit corresponds to user stories 6, 7, 8, 10, 27, 28,
and 29.