Updating the judge controller and mailer
Attempting to fix the judge mailer so that the page will render and load
after an event coordinator creates a judge. In addition I changed the
controller so that it redirects the event coordinator back to the events
page after a judge is created so that they can continue editing/adding
to the event. This commit correlates to user stories 7, 21, and 28.