Mailer SMTP setting change and view layout
Updated the layout for the event page so that you can view and edit the
details of competitors and judges right on the page with "show", "edit",
and "destroy" accordion links and icons (User stories 6, 7, 8, 27, 28,
and 29). In addition, I also updated some of the settings for the SMTP
server for the Judge Mailer (User story 21).