Updating score_template files for nested forms
Increased the functionality within the MVC for the Score Template model
so that we can achieve a sub-nested Score Template form that allows a
user to create a question directly within the Score Template creation
page. More to come soon, hopefully with Ajax requests working within
this form. This corresponds to user story #8