Adding more placeholders and view toying
This commit represents an addition of placeholders within the forms for
signing in an event coordinator or judge (User stories 4 and 12). In
addition, I have been playing around with a lot of view aspects and
variants and have implemented a page transition on signing in a
coordinator just to test and show so that we can get a feel for how we
want page transitions to be implemented (User story 4).