Migration and association
Migrating the database to remove the event id from within the score
template table so that an event only has one score template as the model
suggests (User story 8). In addition, I reflected those changes in the
model as well as in the "form" view as I removed the hidden event_id
field.