jQuery and AJAX functionality
Starting the process of switching necessary functionality over to AJAX
requests for a more efficient and user friendly application, beginning
with competitors. I created the js 'new' and 'create' files in addition
to a js form that allows an AJAX form to be rendered with the proper
jQuery mobile styling for creating a new competitor (User story #10).